requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.0.267
langchain-openai>=0.0.2
langgraph>=0.0.10
openai>=1.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
gspread>=5.7.0
oauth2client>=4.3.3
google-api-python-client>=2.80.0

# Development Dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Note: Make sure to set your OPENAI_API_KEY environment variable
# You can create a .env file in the project root with:
# OPENAI_API_KEY=your-api-key-here
